🎨 Renaming the pages functions more consistently
