+++ pages/start.tsx
@@ -1,13 +1,13 @@
-import styles from "../styles/Home.module.css";
+import styles from "../styles/Start.module.css";
 import Welcome from "../components/messages/Welcome";
 
-export default function Home() {
+export default function Start() {
   return (
     <div className={styles.container}>
       <main className={styles.main}>
         <Welcome />
         <img
-          className={styles.footerimg}
+          className={styles.footerImg}
           src="/needles.svg"
           alt="sewing utensils"
         />
